Move CleanUpEntries to log_processes so we have 1 copy.
Move SignMerkleTree to log_processes so we have 1 copy.
